+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog for rt.ploneversions
 0.9.3 (unreleased)
 ------------------
 
+- Better error message when we cannot read the versions.cfg from dist.plone.org
 - Python 3 support. If you need to use this package on Python2 please use the 0.9.2 version.
   [ale-rt]